+++ IntegrationTests/Areas/Manage/AccountTests.cs
@@ -76,35 +76,6 @@ namespace IntegrationTests.Areas.Manage
             PageSource.ShouldContain(user.Person.Name);
             PageSource.ShouldContain(user.Person.EmailAddress);
         }
-
-        [Fact]
-        public void Create_Role_Test()
-        {
-            username = RandomString();
-            password = RandomString();
-            string roleName = "role_" + RandomString();
-            var user = CreateUser(roles: new string[] { "Access", "Edit", "Admin" });
-            Login();
-
-            Open($"{rootUrl}Lookups/");
-            PageSource.ShouldContain("Lookup Codes");
-
-            Find(text: "Roles").Click();
-            CurrentUrl.ShouldBe($"{rootUrl}Roles");
-
-            Find(css: ".box-tools button[type=submit]").Click();
-            Find(id: "RoleName.NEW").Click();
-            WaitForElement(".editable-input input[type=text]");
-            Find(css: ".editable-input input[type=text]").Clear();
-            Find(css: ".editable-input input[type=text]").SendKeys(roleName);
-            Find(css: ".editable-buttons button[type=submit]").Click();
-            Open($"{rootUrl}Roles");
-            PageSource.ShouldContain(roleName);
-            var adminRole = db.Roles.SingleOrDefault(r => r.RoleName == "Admin");
-            var role = db.Roles.SingleOrDefault(r => r.RoleName == roleName);
-            role.ShouldNotBeNull();
-            role.Priority.GetValueOrDefault().ShouldBeGreaterThan(adminRole.Priority.GetValueOrDefault());
-        }
         
         public void MyData_User_ForgotPassword_Test()
         {
