+++ IntegrationTests/IntegrationTests.csproj
@@ -85,6 +85,7 @@
     <Compile Include="Areas\General\TestScriptController.cs" />
     <Compile Include="Areas\General\GeneralTests.cs" />
     <Compile Include="Areas\Manage\AccountTests.cs" />
+    <Compile Include="Areas\Setup\SetupTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resources\R.Designer.cs">
       <AutoGen>True</AutoGen>
